Rename package to gurobi_modelanalyzer
