--- setup.cfg
@@ -1,7 +1,0 @@
-[options]
-package_dir =
-    =model_analyzer
-packages = find:
-
-[options.packages.find]
-where = model_analyzer
