+++ src/gurobi_modelanalyzer/variable_analyzer.py
@@ -1,6 +1,6 @@
 import gurobipy as gp
 
-import model_analyzer.common as common
+import gurobi_modelanalyzer.common as common
 
 
 def process_variable_types(m, data, objective_variables, constraint_variables):
